ci(build): use jest.config.js if exists

+++ support/test.sh
@@ -4,6 +4,13 @@ PACKAGE_PATH="$SCRIPT_PATH/../packages/$1"
 
 cd "$PACKAGE_PATH"
 
-cp "$ROOT_PATH/jest.config.json" "./"
+JEST_CONFIG_FILE_NAME="$PACKAGE_PATH/jest.config.json"
+JEST_JS_CONFIG_FILE_NAME="$PACKAGE_PATH/jest.config.js"
+if test -f "$JEST_JS_CONFIG_FILE_NAME"
+then
+  JEST_CONFIG_FILE_NAME="$JEST_JS_CONFIG_FILE_NAME"
+else
+  cp "$ROOT_PATH/jest.config.json" "./"
+fi
 
-yarn jest --config ./jest.config.json
+yarn jest --config "$JEST_CONFIG_FILE_NAME"
